extracted example1 details from TestServer
