+++ crabzilla-web/src/test/java/io/github/crabzilla/web/TestServer.kt
@@ -1,20 +1,15 @@
 package io.github.crabzilla.web
 
-import io.github.crabzilla.Snapshot
-import io.github.crabzilla.StateTransitionsTracker
-import io.github.crabzilla.example1.*
-import io.github.crabzilla.pgclient.PgClientEventProjector
 import io.github.crabzilla.pgclient.PgClientUowRepo
-import io.github.crabzilla.pgclient.example1.EXAMPLE1_PROJECTOR_HANDLER
-import io.github.crabzilla.vertx.CommandVerticle
 import io.github.crabzilla.vertx.CrabzillaVerticle
-import io.github.crabzilla.vertx.ProjectionData
 import io.github.crabzilla.vertx.VerticleRole.REST
 import io.github.crabzilla.vertx.initVertx
+import io.github.crabzilla.web.example1.EXAMPLE1_PROJECTION_ENDPOINT
+import io.github.crabzilla.web.example1.customerCmdVerticle
+import io.github.crabzilla.web.example1.setupEventHandler
 import io.reactiverse.pgclient.PgClient
 import io.reactiverse.pgclient.PgPool
 import io.reactiverse.pgclient.PgPoolOptions
-import io.vertx.circuitbreaker.CircuitBreaker
 import io.vertx.config.ConfigRetriever
 import io.vertx.config.ConfigRetrieverOptions
 import io.vertx.config.ConfigStoreOptions
@@ -27,7 +22,6 @@ import io.vertx.core.logging.LoggerFactory.getLogger
 import io.vertx.ext.web.Router
 import io.vertx.ext.web.handler.BodyHandler
 import io.vertx.ext.web.handler.LoggerHandler
-import net.jodah.expiringmap.ExpiringMap
 
 // Convenience method so you can run it in your IDE
 fun main(args: Array<String>) {
@@ -37,8 +31,6 @@ fun main(args: Array<String>) {
 class TestServer(override val name: String = "testServer") : CrabzillaVerticle(name, REST) {
 
   companion object {
-
-    private val PROJECTION_ENDPOINT: String = "example1_projection_endpoint"
 
     internal var httpPort: Int = 8081
     internal var server: HttpServer? = null
@@ -76,13 +68,13 @@ class TestServer(override val name: String = "testServer") : CrabzillaVerticle(n
       initVertx(vertx)
 
       val readDb = pgPool("READ", config)
-      setupEventHandler(readDb)
+      setupEventHandler(vertx, readDb)
       readDbHandle = readDb
 
       val writeDb = pgPool("WRITE", config)
       writeDbHandle = writeDb
       val uowRepository = PgClientUowRepo(writeDb)
-      val customerCmdVerticle = customerCmdVerticle(uowRepository)
+      val customerCmdVerticle = customerCmdVerticle(vertx, uowRepository)
 
       vertx.deployVerticle(customerCmdVerticle)
 
@@ -91,7 +83,7 @@ class TestServer(override val name: String = "testServer") : CrabzillaVerticle(n
       router.route().handler(LoggerHandler.create())
       router.route().handler(BodyHandler.create())
 
-      router.post("/:resource/commands").handler { postCommandHandler(it, uowRepository, PROJECTION_ENDPOINT) }
+      router.post("/:resource/commands").handler { postCommandHandler(it, uowRepository, EXAMPLE1_PROJECTION_ENDPOINT) }
 
       router.get("/:resource/commands/:cmdID").handler { getUowByCmdIdHandler(it, uowRepository) }
 
@@ -129,28 +121,4 @@ class TestServer(override val name: String = "testServer") : CrabzillaVerticle(n
     return PgClient.pool(vertx, writeOptions)
   }
 
-  // related to example1
-
-  fun customerCmdVerticle(uowRepository: PgClientUowRepo): CommandVerticle<Customer> {
-
-    val seedValue = Customer(null, null, false, null, PojoService())
-    val trackerFactory = { snapshot: Snapshot<Customer> -> StateTransitionsTracker(snapshot, CUSTOMER_STATE_BUILDER)}
-    return CommandVerticle("Customer", seedValue, CUSTOMER_CMD_HANDLER, CUSTOMER_CMD_VALIDATOR,
-      trackerFactory, uowRepository, ExpiringMap.create(), CircuitBreaker.create("cb1", vertx))
-
-  }
-
-  fun setupEventHandler(readDb: PgPool) {
-
-    val eventProjector = PgClientEventProjector(readDb, "customer summary")
-
-    vertx.eventBus().consumer<ProjectionData>(PROJECTION_ENDPOINT) { message ->
-
-      println("received events: " + message.body())
-
-      eventProjector.handle(message.body(), EXAMPLE1_PROJECTOR_HANDLER, Future.future())
-
-    }
-
-  }
 }
